Disable process termination in downloader.
It turns out that judicious timeouts can mostly solve that
problem, plus, with it enabled, I was getting a massive amount
of terminated processes on the server. Furthermore, it's just not
a great idea to structure your program around randomly terminating
processes.

Finally, I really need to decrease the amount of process churn,
as it's kind of hard on system resources, and this was the first
step towards doing so.