Add single-file cacheing method.
This should be faster than the previous method with one file per
image, because it shouldn't have to seek all over the disk.

The disadvantage, of course, is decreased flexibility: I had to
go for a worse method of evicting images from the cache, for now.

I also added some unit tests for the cache, which were sorely
overdue.