Add sequential cache loading to ImageGetter.
This is still a very EXPERIMENTAL feature.

Sequential loading is pretty much the fastest way to load images.
For a really large cache, we obviously can load entire batches
this way without really compromising performance. Similarly, for
a really small cache, we clearly don't want to do this, because
we would start seeing the same images over and over again.

The question, then, is what to do for every cache size in between.
The solution implemented here is to use a simple heuristic based
on past cache hit rates in order to determine the number of
additional images to load sequentially on each cache request.